+++ src/main/java/com/bishnet/cucumber/parallel/report/JsonReportMerger.java
@@ -13,23 +13,43 @@ import java.lang.reflect.Type;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class JsonReportMerger {
 
 	private List<Path> reportFiles;
+	private List<Map<String, Object>> features = new ArrayList<>();
+	private Map<String, Map<String, Object>> featuresMap = new HashMap<>();
+	private List<Map<String, Object>> flakyfeatures = new ArrayList<>();
 
 	public JsonReportMerger(List<Path> reportFiles) {
 		this.reportFiles = reportFiles;
 	}
 
 	public void merge(Path mergedReport) throws IOException {
-		List<Map<String, Object>> features = new ArrayList<Map<String, Object>>();
 		for (Path reportFile : reportFiles) {
-			features.addAll(readSingleReport(reportFile));
+			List<Map<String, Object>> featureList = readSingleReport(reportFile);
+			features.addAll(featureList);
 		}
+		writeMergedReportToFile(mergedReport, features);
+	}
 
+	public void mergeRerunFailedReports(Path mergedReport, Path flakyReport) throws IOException {
+		features.addAll(readSingleReport(mergedReport));
+		features.forEach(mergedFeature -> featuresMap.put((String) mergedFeature.get("id"), mergedFeature));
+		readSingleReport(reportFiles.get(0)).forEach(rerunFeature -> {
+			flakyfeatures.add(featuresMap.get((String) rerunFeature.get("id")));
+			featuresMap.put((String) rerunFeature.get("id"), rerunFeature);
+		});
+		features.clear();
+		features.addAll(featuresMap.values());
+		writeMergedReportToFile(flakyReport, flakyfeatures);
+		writeMergedReportToFile(mergedReport, features);
+	}
+
+	private void writeMergedReportToFile(Path mergedReport, List<Map<String, Object>> features) throws IOException {
 		Path reportDirectory = mergedReport.getParent();
 		if (reportDirectory != null)
 			Files.createDirectories(reportDirectory);