+++ src/main/java/com/bishnet/cucumber/parallel/runtime/RuntimeConfiguration.java
@@ -12,13 +12,20 @@ public class RuntimeConfiguration {
 	public final Path htmlReportPath;
 	public final Path jsonReportPath;
 	public final Path threadTimelineReportPath;
+	public final Path rerunReportReportPath;
 	public final boolean htmlReportRequired;
 	public final boolean jsonReportRequired;
 	public final boolean threadTimelineReportRequired;
+	public final boolean rerunReportRequired;
+	public final int flakyAttemptsCount;
+	public final Path flakyReportPath;
+	public final int flakyMaxCount;
 
 	public RuntimeConfiguration(int numberOfThreads, List<String> cucumberPassThroughArguments,
-								List<String> featureParsingArguments, List<String> featurePaths, Path htmlReportPath, boolean htmlReportRequired,
-								Path jsonReportPath, boolean jsonReportRequired, Path threadTimelineReportPath, boolean threadTimelineReportRequired) {
+			List<String> featureParsingArguments, List<String> featurePaths, Path htmlReportPath, boolean htmlReportRequired,
+			Path jsonReportPath, boolean jsonReportRequired, Path threadTimelineReportPath,
+			boolean threadTimelineReportRequired, Path rerunReportReportPath, boolean rerunReportRequired, int flakyAttemptsCount
+			,Path flakyReportPath, int flakyMaxCount) {
 		this.numberOfThreads = numberOfThreads;
 		this.cucumberPassthroughArguments = cucumberPassThroughArguments;
 		this.featureParsingArguments = featureParsingArguments;
@@ -29,6 +36,11 @@ public class RuntimeConfiguration {
 		this.jsonReportRequired = jsonReportRequired;
 		this.threadTimelineReportPath = threadTimelineReportPath;
 		this.threadTimelineReportRequired = threadTimelineReportRequired;
+		this.rerunReportReportPath = rerunReportReportPath;
+		this.rerunReportRequired = rerunReportRequired;
+		this.flakyAttemptsCount = flakyAttemptsCount;
+		this.flakyReportPath = flakyReportPath;
+		this.flakyMaxCount = flakyMaxCount;
 	}
 
 }