+++ src/test/java/com/bishnet/cucumber/parallel/runtime/CucumberRuntimeFactoryTest.java
@@ -1,6 +1,7 @@
 package com.bishnet.cucumber.parallel.runtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,8 +19,21 @@ public class CucumberRuntimeFactoryTest {
 		featurePaths.add("/some/absolute/path");
 		featurePaths.add("a/relative/path");
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
-		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, false);
+		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, Thread.currentThread()
+				.getContextClassLoader());
+		runtimeFactory.getRuntime(cucumberArguments);
+	}
+
+	@Test
+	public void shouldReturnRuntimeWhenPassedValidRerunFile() {
+		List<String> featurePaths = new ArrayList<String>();
+		featurePaths.add("@C:/windows/absolute/path.rerun");
+		featurePaths.add("@a/relative/path.rerun");
+		List<String> cucumberPassthroughArguments = new ArrayList<String>();
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, Thread.currentThread()
 				.getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -32,8 +46,8 @@ public class CucumberRuntimeFactoryTest {
 		featurePaths.add("a/relative/path");
 		featurePaths.add("classpath:");
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
-		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, Thread.currentThread()
 				.getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -44,8 +58,8 @@ public class CucumberRuntimeFactoryTest {
 		List<String> featurePaths = new ArrayList<String>();
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
-		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -57,8 +71,8 @@ public class CucumberRuntimeFactoryTest {
 		List<String> featurePaths = new ArrayList<String>();
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
-		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		assertThat(runtimeFactory.getRuntime(cucumberArguments)).isInstanceOf(Runtime.class);
@@ -69,11 +83,18 @@ public class CucumberRuntimeFactoryTest {
 		List<String> featurePaths = new ArrayList<String>();
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
-		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, true);
+		RuntimeConfiguration runtimeConfiguration = getRuntimeConfiguration(featurePaths,
+				cucumberPassthroughArguments, true);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		assertThat(runtimeFactory.getRuntime(cucumberArguments)).isInstanceOf(ThreadLoggedRuntime.class);
 	}
+
+	private RuntimeConfiguration getRuntimeConfiguration(List<String> featurePaths, List<String>
+			cucumberPassthroughArguments, boolean threadTimelineReportRequired) {
+		return new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
+				null, false, null, false, null, threadTimelineReportRequired, null, false, 0, null, 0);
+	}
+
 
 }