+++ src/test/java/com/bishnet/cucumber/parallel/runtime/FakeCucumberRuntimeFactory.java
@@ -17,7 +17,8 @@ public class FakeCucumberRuntimeFactory extends CucumberRuntimeFactory {
 	private int invocationCount;
 
 	public FakeCucumberRuntimeFactory(byte[] perInvocationExitCodes, boolean[] perInvocationShouldThrowException) {
-		super(new RuntimeConfiguration(0, null, null, null, null, false, null, false, null, false), Thread.currentThread()
+		super(new RuntimeConfiguration(0, null, null, null, null, false, null, false, null, false, null,
+				false, 0, null, 0), Thread.currentThread()
 				.getContextClassLoader());
 		this.perInvocationExitCodes = perInvocationExitCodes;
 		this.perInvocationShouldThrowException = perInvocationShouldThrowException;