+++ src/test/java/com/bishnet/cucumber/parallel/runtime/FeatureSplitterTest.java
@@ -48,6 +48,7 @@ public class FeatureSplitterTest {
 	}
 
 	private RuntimeConfiguration getRuntimeConfiguration(List<String> featureParsingArguments, int numberOfThreads) {
-		return new RuntimeConfiguration(numberOfThreads, null, featureParsingArguments, null, null, false, null, false, null, false);
+		return new RuntimeConfiguration(numberOfThreads, null, featureParsingArguments, null, null, false, null,
+				false, null, false, null, false, 0, null, 0);
 	}
 }