+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ChangeLog
 =====================
 
+0.1.6a (Unreleased)
+=====================
+* Implement rerun failed scenarios + create flaky tests report (cucumber json)
+
+0.1.5 (Unreleased)
+=====================
+* Implement handling for rerun plugin
+
 0.1.4 (2016-11-06)
 =====================
 * Switch to requiring Java 8.